pyproject.toml
[project]
name = "fund-advisor"
version = "0.1.0"
description = "多 Agent CLI 基金推荐程序"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "autogen-agentchat>=0.4",
    "anthropic>=0.40",
    "typer>=0.9",
    "rich>=13.0",
    "akshare>=1.12",
    "tushare>=1.2",
    "jqdatasdk>=1.9",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19",
    "pyinstaller>=6.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "pytest-mock>=3.10",
]

[project.scripts]
fund-advisor = "src.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
